pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "csv2notion"
version = "0.1.3"
description = "Import/Merge CSV files into Notion database"
authors = ["vzhd1701 <vzhd1701@gmail.com>"]
readme = "README.md"
include = ["CHANGELOG.md"]
license = "MIT"
repository = "https://github.com/vzhd1701/csv2notion"
keywords = ["csv", "notion", "import", "merge"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Environment :: Console",
    "Operating System :: MacOS",
    "Operating System :: Microsoft",
    "Operating System :: Unix",
    "Topic :: Internet",
    "Topic :: Terminals",
    "Topic :: Utilities",
]

[tool.poetry.urls]
"Changelog" = "https://github.com/vzhd1701/csv2notion/blob/master/CHANGELOG.md"

[tool.poetry.scripts]
csv2notion = "csv2notion.cli:main"

[tool.poetry.dependencies]
python = "^3.7"
python-dateutil = "^2.8.2"
tqdm = "^4.64.0"
emoji = "^1.7.0"
notion-vzhd1701-fork = "0.0.29"

[tool.poetry.dev-dependencies]
mdformat = "0.7.7"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
pyfakefs = "^4.5.6"
pytest-vcr = "^1.0.2"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.18.1"
wemake-python-styleguide = "^0.16.1"
flakehell = "^0.9.0"
flake8 = "3.9.0"

[tool.flakehell]
base = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/flakehell.toml"
extended_default_ignore=[]
max_line_length = 88
exclude = [
    "tests/*.py",
]

[tool.flakehell.plugins]
# Disabled [personal preference]
# WPS111 short names
# WPS305 forbids f-strings
# WPS326 implicit line concat
# Docstings, don't have any
# WPS226 String constant overuse
# WPS437 protected attributes (needed for Notion sdk)

# Exceptions [false positives & formatting]
# C812 trailing comma <- handled by black
# E203 whitespace before : <- handled by black
# W503 Line break occurred before a binary operator
# WPS317 Indentation for parameters <- handled by black
# WPS323 `%` string formatting, log format

pycodestyle = ["+*", "-E203", "-W503"]
flake8-commas = ["+*", "-C812"]
flake8-quotes = ["+*", "-Q000"]
flake8-rst-docstrings = ["-*"]
flake8-docstrings = ["-*"]
flake8-darglint = ["-*"]
wemake-python-styleguide = ["+*", "-WPS111", "-WPS226", "-WPS305", "-WPS317", "-WPS323", "-WPS326", "-WPS437"]

[tool.flakehell.exceptions."csv2notion/version.py"]
# I like it this way
wemake-python-styleguide = ["-WPS410"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.coverage]
    [tool.coverage.report]
    show_missing = true
    exclude_lines = [
        # Have to re-enable the standard pragma
        'pragma: no cover',

        # Don't complain about missing debug-only code:
        'def __repr__',
        'if self\.debug',

        # Don't complain if tests don't hit defensive assertion code:
        'raise AssertionError',
        'raise NotImplementedError',

        # Don't complain if non-runnable code isn't run:
        'if 0:',
        'if __name__ == .__main__.:'
    ]
    omit = [
        '*/tests/*',
        '*/notion/*',
    ]
